3.5-KOD-Wprowadzenie do testów jednostkowych-ZADANIE

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/ResultChecker.java
@@ -2,6 +2,6 @@ package com.kodilla.basic_assertion;
 
 public class ResultChecker {
 
-    public static boolean assertEquals(int expected, int actual) {
+    public static boolean assertEquals(double expected, double actual) {
         return expected == actual;
     }}
